update the unittest for Common.h of libconsensus
