+++ test/unittests/libtxpool/FakeBlockChain.h
@@ -26,6 +26,7 @@
 #include <libtxpool/TxPool.h>
 #include <test/tools/libutils/TestOutputHelper.h>
 #include <test/unittests/libethcore/FakeBlock.h>
+#include <test/unittests/libp2p/FakeHost.h>
 #include <boost/test/unit_test.hpp>
 using namespace dev;
 using namespace dev::txpool;
@@ -99,6 +100,44 @@ public:
     std::vector<std::shared_ptr<Block> > m_blockChain;
     uint64_t m_blockNumber;
 };
+class TxPoolFixture
+{
+public:
+    TxPoolFixture(uint64_t blockNum = 5, size_t const& transSize = 5)
+    {
+        FakeTxPoolFunc(blockNum, transSize);
+    }
 
+    void FakeTxPoolFunc(uint64_t _blockNum, size_t const& transSize)
+    {
+        /// fake block manager
+        m_blockChain = std::make_shared<FakeBlockChain>(_blockNum, transSize);
+        /// fake host of p2p module
+        FakeHost* hostPtr = createFakeHostWithSession(clientVersion, listenIp, listenPort);
+        m_host = std::shared_ptr<Host>(hostPtr);
+        /// create p2pHandler
+        m_p2pHandler = std::make_shared<P2PMsgHandler>();
+        /// fake service of p2p module
+        m_topicService = std::make_shared<Service>(m_host, m_p2pHandler);
+        std::shared_ptr<BlockChainInterface> blockChain =
+            std::shared_ptr<BlockChainInterface>(m_blockChain);
+        /// fake txpool
+        m_txPool = std::make_shared<FakeTxPool>(m_topicService, blockChain);
+    }
+
+    void setSessionData(std::string const& data_content)
+    {
+        for (auto session : m_host->sessions())
+            dynamic_cast<FakeSessionForTest*>(session.second.get())->setDataContent(data_content);
+    }
+    std::shared_ptr<FakeTxPool> m_txPool;
+    std::shared_ptr<Service> m_topicService;
+    std::shared_ptr<P2PMsgHandler> m_p2pHandler;
+    std::shared_ptr<FakeBlockChain> m_blockChain;
+    std::shared_ptr<Host> m_host;
+    std::string clientVersion = "2.0";
+    std::string listenIp = "127.0.0.1";
+    uint16_t listenPort = 30304;
+};
 }  // namespace test
 }  // namespace dev
